+++ src/components/UI/Records/features/TransferAccountSelector/SelectDestinationAccount.tsx
@@ -1,17 +1,16 @@
 import { SelectChangeEvent } from '@mui/material';
 import { Select } from '../../../../../styles';
-import { SelectOriginAccountProps } from './TransferAccountSelector.interfaces';
+import { SelectDestinationAccountProps } from './TransferAccountSelector.interfaces';
 
-const SelectOriginAccount = ({
-  children, field, form, selectOriginAccount,
-}: SelectOriginAccountProps) => {
+const SelectDestinationAccount = ({
+  children, field, form, toggleDestinationCredit,
+}: SelectDestinationAccountProps) => {
   const { name, value } = field;
   const { setFieldValue } = form;
 
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const handleChange = (event: SelectChangeEvent<any>) => {
-    selectOriginAccount(event.target.value);
-    setFieldValue('destinationAccount', '');
+    toggleDestinationCredit(event.target.value);
     setFieldValue(name, event.target.value);
   };
 
@@ -26,4 +25,4 @@ const SelectOriginAccount = ({
   );
 };
 
-export { SelectOriginAccount };
+export { SelectDestinationAccount };
